+++ jsk_recognition/package.xml
@@ -24,6 +24,7 @@
   <run_depend>jsk_perception</run_depend>
   <run_depend>resized_image_transport</run_depend>
   <run_depend>jsk_recognition_msgs</run_depend>
+  <run_depend>jsk_recognition_utils</run_depend>
   <export>
     <metapackage/>
   </export>
