+++ jsk_recognition_utils/include/jsk_recognition_utils/tf_listener_singleton.h
@@ -33,49 +33,34 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  *********************************************************************/
 
-#include <jsk_topic_tools/log_utils.h>
-#include "jsk_pcl_ros/tf_listener_singleton.h"
-#include <boost/format.hpp>
 
-namespace jsk_pcl_ros
+#ifndef JSK_RECOGNITION_UTILS_TF_LISTENER_SINGLETON_H_
+#define JSK_RECOGNITION_UTILS_TF_LISTENER_SINGLETON_H_
+
+#include <tf/transform_listener.h>
+
+namespace jsk_recognition_utils
 {
-  tf::TransformListener* TfListenerSingleton::getInstance()
+  class TfListenerSingleton
   {
-    boost::mutex::scoped_lock lock(mutex_);
-    if (!instance_) {
-      JSK_ROS_INFO("instantiating tf::TransformListener");
-      instance_ = new tf::TransformListener(ros::Duration(30.0));
-    }
-    return instance_;
-  }
+  public:
+    static tf::TransformListener* getInstance();
+    static void destroy();
+  protected:
+    static tf::TransformListener* instance_;
+    static boost::mutex mutex_;
+  private:
+    TfListenerSingleton(TfListenerSingleton const&){};
+    TfListenerSingleton& operator=(TfListenerSingleton const&){};
+  };
 
-  void TfListenerSingleton::destroy()
-  {
-    boost::mutex::scoped_lock lock(mutex_);
-    if (instance_) {
-      delete instance_;
-    }
-  }
-
+  // tf Utility
   tf::StampedTransform lookupTransformWithDuration(
     tf::TransformListener* listener,
     const std::string& to_frame,
     const std::string& from_frame,
     const ros::Time& stamp,
-    ros::Duration duration)
-  {
-    if (listener->waitForTransform(from_frame, to_frame, stamp, duration)) {
-      tf::StampedTransform transform;
-      listener->lookupTransform(
-        from_frame, to_frame, stamp, transform);
-      return transform;
-    }
-    throw tf2::TransformException(
-      (boost::format("Failed to lookup transformation from %s to %s")
-       % from_frame.c_str() % to_frame.c_str()).str().c_str());
-      
-  }
-  
-  tf::TransformListener* TfListenerSingleton::instance_;
-  boost::mutex TfListenerSingleton::mutex_;
+    ros::Duration duration);
 }
+
+#endif
